# Significantly improve tight layout performance in presence of cartopy axes