--- a/lib/cartopy/mpl/geoaxes.py
+++ b/lib/cartopy/mpl/geoaxes.py
@@ -494,16 +494,17 @@ class GeoAxes(matplotlib.axes.Axes):
     def get_default_bbox_extra_artists(self):
         """
         Return a default list of artists that are used for the bounding box
-        calculation. Artists are excluded by not being visible, by being clipped
-        by the axes patch path, or by ``artist.set_in_layout(False)``.
+        calculation. Artists are excluded by not being visible, by being
+        clipped by the axes patch path, or by ``artist.set_in_layout(False)``.
         """
         artists_keep = (
             matplotlib.axes.Axes, matplotlib.axis.Axis, matplotlib.spines.Spine
         )
+        transform = mtransforms.TransformedPatchPath
         artists = [
             artist for artist in super().get_default_bbox_extra_artists()
             if isinstance(artist, artists_keep)
-            or not isinstance(artist.get_clip_path(), mtransforms.TransformedPatchPath)
+            or not isinstance(artist.get_clip_path(), transform)
             or artist.get_clip_path()._patch is not self.patch
         ]
         return artists