@lukelbd 2021-12-04
This is also one of those PRs that might require a thousand test image replacements, because I'm guessing "clipped" artists were probably effecting the tight layout by a pixel or two...... ugh.