@dopplershift 2021-12-04
This sounds like a nice improvement! I'm split on whether this belongs here or in matplotlib, since this isn't using anything specific to us, it's just a case that's particularly important here.

@tacaswell @jklymak thoughts?