@jklymak 2021-12-05
`TransformedPatchPath` seems a pretty general check.  Is there not an axes object that can be tested instead?