@lukelbd 2021-12-05
Just to be clear I'm checking the `ax.patch` (of type `geoaxes._ViewClippedPatch`) from which the `clip_path` (of type `TransformedPatchPath`) is derived (see [the `set_clip_path` code](https://github.com/matplotlib/matplotlib/blob/c0fc78160a97264792dbb2eba964ad6975074614/lib/matplotlib/artist.py#L759-L819)). When `TransformedPatchPath` is instantiated, the parent patch is stored as a `_patch` attribute so that it can be automatically updated (see [the `TransformedPatchPath` code](https://github.com/matplotlib/matplotlib/blob/c0fc78160a97264792dbb2eba964ad6975074614/lib/matplotlib/transforms.py#L2799-L2835)). Note that the private `_patch` attribute [dates back to the origin](https://github.com/matplotlib/matplotlib/commit/8357131482fcdfff307a9846128ebbb5e9df7c03) of `TransformedPatchPath` in matplotlib 2.1. 

Re: Putting this in matplotlib vs. cartopy, on reflection, I think there are a few advantages to putting it here:

* It looks like [`set_clip_path`](https://github.com/matplotlib/matplotlib/blob/c0fc78160a97264792dbb2eba964ad6975074614/lib/matplotlib/artist.py#L759-L819) accepts a variety of input types -- not just `Patch` instances. A PR submitted to matplotlib might incur requests for a less "obvious", more general solution appopriate for all input types.
* Matplotlib's [`get_tightbbox`](https://github.com/matplotlib/matplotlib/blob/9765379ce6e7343070e815afc0988874041b98e2/lib/matplotlib/axes/_base.py#L4646-L4657) bypasses artists when the `Bbox.extents` of its `clip_box` are identical to `ax.bbox.extents`. Again, a PR submitted to matplotlib might incur requests for a similarly general solution comparing the actual `clip_path` coordinates to `TransformedPatchPath(ax.patch)` rather than just checking the `_patch` identity... which could involve more design choices and possibly reduce performance gains.
* Most importantly, I'm not sure whether "the axes patch edge defines the subplot edge" is standard procedure across all matplotlib-dependent projects. Another library might use larger or smaller axes patches that are not necessarily synonymous with the subplot edge. This solution is "stable" for cartopy because cartopy does (?) use the axes patch in this way, but I'm guessing the choice isn't canon.

A last point (of course) is that submitting here makes the improvement backwards-compatible with older matplotlib versions. Which seems like a practical advantage.

One thing that bothers me is I have no idea *where* matplotlib artists get clipped by `clip_box`, or why cartopy artists do not. I recursive grep inside matplotlib shows nothing obvious in the python code:

<details>
<summary>Grep results</summary>

```python
$ grep -r 'set_clip_box' .
../legend_handler.py:        legend_handle.set_clip_box(None)
../legend_handler.py:        legend_handle.set_clip_box(None)
../axes/_base.py:            _title.set_clip_box(None)
../tests/test_image.py:    im.set_clip_box(axs[0].bbox)
../tests/test_polar.py:    el.set_clip_box(ax.bbox)
../artist.py:    def set_clip_box(self, clipbox):
../text.py:    def set_clip_box(self, clipbox):
../text.py:        super().set_clip_box(clipbox)
../contour.py:        label.set_clip_box(self.axes.bbox)
../patheffects.py:        self.patch.set_clip_box(gc.get_clip_rectangle())
```

</details>

By contrast, a recursive grep of `set_clip_path` indicates that *all* artists (not just cartopy artists) are clipped by `ax.patch`:

<details>
<summary>Grep results</summary>

```python
$ grep -r 'set_clip_path' .
../legend_handler.py:        legend_handle.set_clip_path(None)
../legend_handler.py:        legend_handle.set_clip_path(None)
../axis.py:    def set_clip_path(self, clippath, transform=None):
../axis.py:        super().set_clip_path(clippath, transform)
../axis.py:        self.gridline.set_clip_path(clippath, transform)
../axis.py:            self.set_clip_path(self.axes.patch)
../axis.py:    def set_clip_path(self, clippath, transform=None):
../axis.py:        super().set_clip_path(clippath, transform)
../axis.py:            child.set_clip_path(clippath, transform)
../backend_bases.py:    def set_clip_path(self, path):
../axes/_axes.py:        t.set_clip_path(self.patch)
../axes/_axes.py:            a.set_clip_path(self.patch)
../axes/_axes.py:            line.set_clip_path(self.patch)
../axes/_axes.py:            im.set_clip_path(self.patch)
../axes/_axes.py:            ret.set_clip_path(self.patch)
../axes/_base.py:        self.xaxis.set_clip_path(self.patch)
../axes/_base.py:        self.yaxis.set_clip_path(self.patch)
../axes/_base.py:        a.set_clip_path(self.patch)
../axes/_base.py:            collection.set_clip_path(self.patch)
../axes/_base.py:            line.set_clip_path(self.patch)
../axes/_base.py:            p.set_clip_path(self.patch)
../axes/_base.py:        tab.set_clip_path(self.patch)
../colorbar.py:        col.set_clip_path(mpath.Path(xy, closed=True),
../backends/backend_cairo.py:    def set_clip_path(self, path):
../tests/test_artist.py:    col.set_clip_path(clip_path, ax1.transData)
../tests/test_artist.py:    patch.set_clip_path(clip_path, ax2.transData)
../tests/test_artist.py:    l.set_clip_path(p)
../tests/test_image.py:    im.set_clip_path(patch)
../tests/test_image.py:    im.set_clip_path(rect)
../tests/test_image.py:    im.set_clip_path(None)
../tests/test_bbox_tight.py:    patch.set_clip_path(path, transform=ax.transAxes)
../figure.py:            artist.set_clip_path(self.patch)
../offsetbox.py:                c.set_clip_path(tpath)
../projections/polar.py:            self.yaxis.set_clip_path(self.patch)
../artist.py:    def set_clip_path(self, path, transform=None):
../artist.py:                "Invalid arguments to set_clip_path, of type {} and {}"
../artist.py:            gc.set_clip_path(self._clippath)
../artist.py:            gc.set_clip_path(None)
../text.py:    def set_clip_path(self, path, transform=None):
../text.py:        super().set_clip_path(path, transform)
../image.py:                    gc.set_clip_path(parent.get_clip_path())
../patheffects.py:            self.patch.set_clip_path(*clip_path)
```

</details>

So, this PR "works" because cartopy artists are not clipped by `clip_box`, not because *only* cartopy artists are clipped by `ax.patch`. Curious if anyone knows why/where `clip_box` is skipped for cartopy artists (though maybe it's not critical).