@greglucas 2021-12-05
Does this only help with pcolormesh? My initial guess is that most of the artists that are causing the slowdown are the pcolor cells involved in the wrap, which can be found with `mesh._wrapped_collection_fix`, so I wonder if we aren't adding those to the axes properly or need to do some additional work there?

https://github.com/SciTools/cartopy/blob/d35c910cbd1ee8a38e8395e0a00ff06963435448/lib/cartopy/mpl/geoaxes.py#L1900-L1920

If we remove these wrapped cells from the bbox calculation I'm wondering if we would start to shrink the limits when the cells are wrapped by not accounting for the pcolor anymore. Basically what I'm getting at is that I _think_ the slowdown is due to slow trans_non_affine calculations from the pcolor cell bounds calculation and we are just removing that slow calculation here, but at the loss of auto-scaling bounds calculations.